+++ src/main/kotlin/com/sight/controllers/http/PingController.kt
@@ -1,4 +1,4 @@
-package com.sight.controllers.api
+package com.sight.controllers.http
 
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
