+++ src/main/kotlin/com/sight/controllers/http/TestController.kt
@@ -1,8 +1,8 @@
-package com.sight.controllers.api
+package com.sight.controllers.http
 
-import com.sight.core.Auth
-import com.sight.domain.auth.Requester
-import com.sight.domain.auth.UserRole
+import com.sight.core.auth.Auth
+import com.sight.core.auth.Requester
+import com.sight.core.auth.UserRole
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
