+++ src/main/kotlin/com/sight/repository/DiscordIntegrationRepository.kt
@@ -1,6 +1,6 @@
 package com.sight.repository
 
-import com.sight.domain.DiscordIntegration
+import com.sight.domain.discord.DiscordIntegration
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
